Use PrimMonad to make life easier
* As usual, the `IO` and `ST` code were actually the same.
`PrimMonad` abstracts from that, and also adds support for
monad transformers.

* Removed utterly bogus `Show` instances for monadic graphs.
  The right thing to do is probably to offer a graph freezing
  primitive; then the frozen graph can be shown as usual. For
  large graphs, streaming conversion would be useful.
